src/teacher_web/requirements.txt
django
django-debug-toolbar
django-cors-headers
mysqlclient
mysql-connector-python
djangorestframework
requests
python-dotenv

